+++ Extension/src/pages/filtering-log/components/FilteringEvents/FilteringEvents.jsx
@@ -213,6 +213,7 @@ const statusAccessor = (props) => {
 };
 
 const Row = observer(({
+    index,
     event,
     columns,
     onClick,
@@ -227,15 +228,20 @@ const Row = observer(({
     );
 
     return (
-        <button
+        // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/interactive-supports-focus
+        <div
+            role="row"
             style={{
                 ...style,
                 top: `${parseFloat(style.top) + ITEM_HEIGHT_PX}px`,
             }}
             id={event.eventId}
             onClick={onClick}
-            type="button"
             className={className}
+            // Set row index explicitly for screen readers, because table
+            // is virtualized and not all of the rows are rendered at the same time.
+            // Add 2 to index to include the header row and make it 1-based index.
+            aria-rowindex={index + 2}
         >
             {
                 columns.map((column) => {
@@ -249,6 +255,7 @@ const Row = observer(({
 
                     return (
                         <div
+                            role="cell"
                             className="td"
                             key={column.id}
                             style={{ width: column.getWidth() }}
@@ -258,7 +265,14 @@ const Row = observer(({
                     );
                 })
             }
-        </button>
+            {/* This cell is available only for screen readers to notify
+                users that they can open details of this log with keyboard. */}
+            <div role="cell" className="sr-only">
+                <button type="button" onClick={onClick}>
+                    {translator.getMessage('filtering_table_open_details')}
+                </button>
+            </div>
+        </div>
     );
 });
 
@@ -272,6 +286,7 @@ const VirtualizedRow = ({
 
     return (
         <Row
+            index={index}
             event={event}
             columns={columns}
             onClick={handleRowClick}
@@ -285,24 +300,36 @@ const ColumnsContext = React.createContext({});
 const ColumnsProvider = ColumnsContext.Provider;
 
 const TableHeader = ({ style }) => {
-    const { columns } = useContext(ColumnsContext);
+    const { columns, isLogEventsEmpty } = useContext(ColumnsContext);
 
+    /**
+     * WAI ARIA attributes are hidden if the table is empty, this needed
+     * to properly announce FilteringEventsEmpty component, otherwise
+     * screen readers will ignore this block.
+     */
     return (
         <div
+            role="row"
             className="thead"
             style={style}
+            aria-hidden={isLogEventsEmpty}
+            // Set row index explicitly for screen readers, because table
+            // is virtualized and not all of the rows are rendered at the same time.
+            // Header row is always first row in the table (1-based index).
+            aria-rowindex={isLogEventsEmpty ? undefined : 1}
         >
             <div className="tr">
                 {
                     columns.map((column) => (
                         <div
+                            role="columnheader"
                             className="th"
                             key={column.id}
                             style={{ width: column.getWidth() }}
                         >
                             {column.Header}
                             <div
-                                role="separator"
+                                aria-hidden="true"
                                 className="resizer"
                                 key={column.id}
                                 style={{ cursor: 'col-resize' }}
@@ -311,6 +338,11 @@ const TableHeader = ({ style }) => {
                         </div>
                     ))
                 }
+                {/* This column is available only for screen readers to notify
+                    users that they can open details of this log with keyboard. */}
+                <div role="columnheader" className="sr-only">
+                    {translator.getMessage('filtering_table_action')}
+                </div>
             </div>
         </div>
     );
@@ -338,6 +370,7 @@ const FilteringEventsRows = observer(({
     handleRowClick,
 }) => {
     const { events } = logStore;
+    const isLogEventsEmpty = logStore.events.length === 0;
 
     return (
         /**
@@ -346,7 +379,7 @@ const FilteringEventsRows = observer(({
          *
          * @see {@link https://github.com/bvaughn/react-window/issues/404}
          */
-        <ColumnsProvider value={{ columns }}>
+        <ColumnsProvider value={{ columns, isLogEventsEmpty }}>
             <AutoSizer>
                 {({
                     height,
@@ -378,6 +411,8 @@ const MIN_COLUMN_WIDTH = 50;
 
 const FilteringEvents = observer(() => {
     const { logStore } = useContext(rootStore);
+
+    const isLogEventsEmpty = logStore.events.length === 0;
 
     const tableRef = useRef(null);
 
@@ -571,8 +606,21 @@ const FilteringEvents = observer(() => {
 
     const columns = addMethods(columnsData);
 
+    /**
+     * WAI ARIA attributes are hidden if the table is empty, this needed
+     * to properly announce FilteringEventsEmpty component, otherwise
+     * screen readers will ignore this block.
+     */
     return (
-        <div className="filtering-log">
+        <div
+            role={isLogEventsEmpty ? undefined : 'table'}
+            className="filtering-log"
+            aria-label={isLogEventsEmpty ? undefined : translator.getMessage('filtering_log_title')}
+            // Set number of rows explicitly for screen readers, because table
+            // is virtualized and not all of the rows are rendered at the same time.
+            // Add 1 to the number of rows to include the header row.
+            aria-rowcount={isLogEventsEmpty ? undefined : logStore.events.length + 1}
+        >
             <div
                 style={{ minWidth: `${minTableWidth}px` }}
                 className="table filtering-log__inner"