+++ Extension/src/pages/filtering-log/components/Status/Status.jsx
@@ -53,10 +53,33 @@ export const Status = (props) => {
     const isModified = mode === StatusMode.MODIFIED;
     const areNetworkBadgesVisible = requestUrl && !isModified;
     const statusTooltipText = getStatusTitle(mode);
+    const statusToShow = statusCode || '----';
+
+    // This text is rendered only for Screen readers as summary of the status
+    let requestStatusSummary = `${timeString}.`;
+
+    if (areNetworkBadgesVisible) {
+        requestStatusSummary += ` ${translator.getMessage('filtering_log_tag_request_status')}: ${statusTooltipText}.`;
+
+        if (!isBlocked) {
+            requestStatusSummary += ` ${translator.getMessage('filtering_log_badge_tooltip_http_status_code')}: ${statusToShow}.`;
+        }
+    }
+
+    if (method) {
+        requestStatusSummary += ` ${translator.getMessage('filtering_log_badge_tooltip_http_req_method')}: ${method}.`;
+    }
+
+    if (requestThirdParty) {
+        requestStatusSummary += ` ${translator.getMessage('filtering_log_badge_tooltip_third_party')}.`;
+    }
 
     return (
         <div className="status-wrapper">
-            <div className="status">
+            <span className="sr-only">
+                {requestStatusSummary}
+            </span>
+            <div className="status" aria-hidden="true">
                 {/* Time string may have different width
                     Preventing layout shift with fixed value
                 */}
@@ -71,6 +94,7 @@ export const Status = (props) => {
                                     <Icon
                                         id={statusCode ? '#transfer-status' : '#arrow-status'}
                                         classname="status__icon"
+                                        aria-hidden="true"
                                     />
                                 </div>
                             </Popover>
@@ -79,13 +103,17 @@ export const Status = (props) => {
                             {isBlocked ? (
                                 <Popover text={translator.getMessage('filtering_log_status_blocked')}>
                                     <div className="status__icon--wrapper">
-                                        <Icon id="#ban" classname="status__icon" />
+                                        <Icon
+                                            id="#ban"
+                                            classname="status__icon"
+                                            aria-hidden="true"
+                                        />
                                     </div>
                                 </Popover>
                             ) : (
                                 <Popover text={translator.getMessage('filtering_log_badge_tooltip_http_status_code')}>
                                     <div className={badgeClassNames}>
-                                        {statusCode || '----'}
+                                        {statusToShow}
                                     </div>
                                 </Popover>
                             )}