+++ Extension/src/pages/options/components/Filters/Group.jsx
@@ -76,6 +76,8 @@ const renderEnabledFilters = (enabledFilters) => {
 };
 
 const DisabledCustomFiltersGroup = ({
+    titleId,
+    descriptionId,
     groupName,
     groupId,
     enabledFilters,
@@ -86,6 +88,8 @@ const DisabledCustomFiltersGroup = ({
         'group--disabled': true,
     });
 
+    const warningDescriptionId = `${descriptionId}-warning`;
+
     return (
         <div className={groupClassName}>
             <a
@@ -93,27 +97,31 @@ const DisabledCustomFiltersGroup = ({
                 target="_blank"
                 rel="noreferrer"
                 className="setting__area setting__area_group link_like_setting_area"
+                aria-labelledby={titleId}
+                aria-describedby={`${descriptionId} ${warningDescriptionId}`}
             >
-                <div className="container">
+                <span className="container">
                     <Icon
                         id={`#setting-${groupId}`}
                         classname="icon--24 setting__icon"
+                        aria-hidden="true"
                     />
-                    <div className="setting__info">
-                        <div className="setting__title group__title">
+                    <span className="setting__info">
+                        <span id={titleId} className="setting__title group__title">
                             {groupName}
-                        </div>
-                        <div className="setting__desc">
+                        </span>
+                        <span id={descriptionId} className="setting__desc">
                             {renderEnabledFilters(enabledFilters)}
-                        </div>
-                        <span className="warning">
+                        </span>
+                        <span id={warningDescriptionId} className="warning">
                             {translator.getMessage('options_filters_custom_disabled_cws')}
                         </span>
-                    </div>
-                </div>
+                    </span>
+                </span>
                 <Icon
                     id="#link"
                     classname="icon icon--24 icon--green-default link"
+                    aria-hidden="true"
                 />
             </a>
         </div>
@@ -134,10 +142,16 @@ const Group = ({
         'group--disabled': !checkboxValue,
     });
 
+    const titleId = `setting-title-${groupId}`;
+    const descriptionId = `setting-desc-${groupId}`;
+    const iconId = `#setting-${groupId}`;
+
     // TODO: Remove this component when custom filters will be supported for MV3.
     if (__IS_MV3__ && groupId === AntibannerGroupsId.CustomFiltersGroupId) {
         return (
             <DisabledCustomFiltersGroup
+                titleId={titleId}
+                descriptionId={descriptionId}
                 groupName={groupName}
                 groupId={AntibannerGroupsId.CustomFiltersGroupId}
                 enabledFilters={enabledFilters}
@@ -146,38 +160,43 @@ const Group = ({
     }
 
     return (
-        <div className={groupClassName}>
+        <li className={groupClassName}>
             <button
                 type="button"
+                role="link"
                 tabIndex={0}
                 className="setting__area setting__area_group"
                 onClick={groupClickHandler}
+                aria-labelledby={titleId}
+                aria-describedby={descriptionId}
             >
                 <Icon
-                    id={`#setting-${groupId}`}
+                    id={iconId}
                     classname="icon--24 setting__icon"
+                    aria-hidden="true"
                 />
-                <div className="setting__info">
-                    <div className="setting__title group__title">
+                <span className="setting__info">
+                    <span id={titleId} className="setting__title group__title">
                         {groupName}
-                    </div>
-                    <div className="setting__desc">
+                    </span>
+                    <span id={descriptionId} className="setting__desc">
                         {renderEnabledFilters(enabledFilters)}
-                    </div>
-                </div>
+                    </span>
+                </span>
             </button>
             <div className="setting__inline-control setting__inline-control_group">
                 <Setting
                     id={groupId}
                     type={SETTINGS_TYPES.CHECKBOX}
                     label={groupName}
+                    labelId={titleId}
                     value={checkboxValue}
                     handler={checkboxHandler}
                     optimistic={!__IS_MV3__}
                     className="group__checkbox"
                 />
             </div>
-        </div>
+        </li>
     );
 };