+++ Extension/src/pages/options/components/Sidebar/Sidebar.jsx
@@ -16,7 +16,11 @@
  * along with AdGuard Browser Extension. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { useContext, useState } from 'react';
+import React, {
+    useContext,
+    useLayoutEffect,
+    useState,
+} from 'react';
 import { observer } from 'mobx-react';
 
 import classNames from 'classnames';
@@ -25,18 +29,51 @@ import { Icon } from '../../../common/components/ui/Icon';
 import { rootStore } from '../../stores/RootStore';
 import { Nav } from '../Nav';
 import { messenger } from '../../../services/messenger';
+import { translator } from '../../../../common/translators/translator';
 
 import { Compare } from './Compare';
 
 import './sidebar.pcss';
 
+const SIDEBAR_ID = 'sidebar';
+
 const Sidebar = observer(() => {
-    const { settingsStore } = useContext(rootStore);
+    const { settingsStore, uiStore } = useContext(rootStore);
 
-    const [isOpen, setOpen] = useState(false);
+    const { isSidebarOpen, openSidebar, closeSidebar } = uiStore;
 
-    const openSidebar = () => setOpen(true);
-    const closeSidebar = () => setOpen(false);
+    const tabletAndAboveQuery = '(min-width: 640px)';
+    const [isTabletAndAboveScreen, setIsTabletAndAboveScreen] = useState(
+        window.matchMedia(tabletAndAboveQuery).matches,
+    );
+
+    useLayoutEffect(() => {
+        const matchMedia = window.matchMedia(tabletAndAboveQuery);
+
+        const handleScreenChange = (e) => {
+            setIsTabletAndAboveScreen(e.matches);
+        };
+
+        // Triggered at the first client-side load and if query changes
+        setIsTabletAndAboveScreen(matchMedia.matches);
+
+        matchMedia.addEventListener('change', handleScreenChange);
+
+        return () => {
+            matchMedia.removeEventListener('change', handleScreenChange);
+        };
+    }, []);
+
+    // Lock sidebar from tab focus if sidebar is closed and it's below tablet screen size.
+    const isSidebarLocked = !isSidebarOpen && !isTabletAndAboveScreen;
+
+    const toggleSidebar = () => {
+        if (isSidebarOpen) {
+            closeSidebar();
+        } else {
+            openSidebar();
+        }
+    };
 
     /**
      * Close sidebar and remove specific limit warning, e.g. on reaching dynamic rules limit on custom filter enabling.
@@ -55,35 +92,48 @@ const Sidebar = observer(() => {
 
     const className = classNames('sidebar', {
         /* styles only for mobile markup */
-        'sidebar--open': isOpen,
+        'sidebar--open': isSidebarOpen,
+    });
+    const overlayClassName = classNames('sidebar__overlay', {
+        'sidebar__overlay--visible': isSidebarOpen,
     });
 
     return (
         <>
-            {isOpen
-                ? (
-                    <div
-                        role="menu"
-                        tabIndex={0}
-                        onClick={closeSidebarWrapper}
-                        onKeyUp={closeSidebarWrapper}
-                        className="sidebar__overlay"
+            <div className="sidebar__menu">
+                <button
+                    type="button"
+                    className="sidebar__menu-button"
+                    aria-label={translator.getMessage('options_navigation')}
+                    aria-expanded={isSidebarOpen}
+                    aria-controls={SIDEBAR_ID}
+                    onClick={toggleSidebar}
+                >
+                    <Icon
+                        id={isSidebarOpen ? '#cross' : '#menu'}
+                        classname={isSidebarOpen ? 'icon--24 icon--gray-default' : 'icon--menu'}
+                        aria-hidden="true"
                     />
-                )
-                : (
-                    <div className="sidebar__menu" role="menu">
-                        <button onClick={openSidebar} className="sidebar__open-button" type="button">
-                            <Icon id="#menu" classname="icon--menu" />
-                        </button>
-                    </div>
-                )}
-            <div className={className}>
-                <Icon id="#logo" classname="icon--logo sidebar__logo" />
-                <Nav closeSidebar={closeSidebarWrapper} />
+                </button>
+            </div>
+            {/* eslint-disable-next-line max-len */}
+            {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */}
+            <div className={overlayClassName} onClick={closeSidebar} />
+            <div
+                id={SIDEBAR_ID}
+                className={className}
+                inert={isSidebarLocked ? '' : undefined}
+            >
+                <Icon
+                    id="#logo"
+                    classname="icon--logo sidebar__logo"
+                    aria-hidden="true"
+                />
+                <Nav onLinkClick={closeSidebarWrapper} />
                 {settingsStore.showAdguardPromoInfo && (
                     <Compare
-                        click={handleCompareClick}
-                        hide={hideCompare}
+                        onCompareClick={handleCompareClick}
+                        onCloseClick={hideCompare}
                     />
                 )}
             </div>