+++ Extension/src/pages/popup/components/Actions/SingleActions/ReportIssueAction.tsx
@@ -29,6 +29,8 @@ import { type SingleActionParams } from './types';
 import '../actions.pcss';
 
 export const ReportIssueAction = ({ className, isFilteringPossible, url }: SingleActionParams) => {
+    const title = translator.getMessage('popup_abuse_site');
+
     /**
      * Handle issue reporting action click.
      */
@@ -50,15 +52,17 @@ export const ReportIssueAction = ({ className, isFilteringPossible, url }: Singl
         <button
             type="button"
             className={className}
+            title={title}
             onClick={handleAbuseSite}
         >
             <Icon
                 id="#thumb-down"
                 classname="icon--24 icon--action--thumb-down"
+                aria-hidden="true"
             />
-            <div className="action__title">
-                {translator.getMessage('popup_abuse_site')}
-            </div>
+            <span className="action__title">
+                {title}
+            </span>
         </button>
     );
 };